requirements.txt
torch>=2.0.0
h5py>=2.10.0
numpy>=1.19.0
pyyaml>=5.4.1 
pandas>=1.3.0
scikit-learn>=1.3.0
wandb>=0.15.0
python-dotenv>=1.0.1
torchsummary>=1.5.1
matplotlib>=3.5.0
seaborn>=0.12.0
statsmodels>=0.13.0
